+++ yodeploy/test_integration/samples/nginx-app/scripts/build.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+set -eu
+cd "$(dirname "$0")/.."
+
+rm -Rf build
+mkdir build
+cp -R deploy build
+cp -R app build